Listen to web view loading progress

+++ Sources/core-web-browser/Window/WindowPresentableModel.swift
@@ -4,5 +4,5 @@ public struct WindowPresentableModel {
     public let showWebView: Bool
     public let canGoBack: Bool
     public let canGoForward: Bool
-    public let progressBar: Int?
+    public let progressBar: Double?
 }
